sympify: losslessly convert strings to high-precision floats
Previously, the output of `sympify(<str>)` always gave precision
of 15, even if the string was much longer.

Now, when `sympify` (actually auto_number() in sympy_parser.py) calls
`Float` it now calls it as `Float(<str>, "")` where the null string
automatically adjusts the precision.

A consequence is that the output has various precisions.  E.g.,
`sympify('1.24')` now has a precision of 3 (rather than 15 as before).

Fixes #8821.